Fix OSD time and e.g. frame number displayed with osd_fractions getting out of sync by using the same source for both.
Patch by Roger Pack [rogerdpack2 gmail com].


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34400 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demuxer.c
@@ -1480,7 +1480,7 @@ double demuxer_get_time_length(demuxer_t *demuxer)
  *        0 otherwise
  * \return the current play time
  */
-int demuxer_get_current_time(demuxer_t *demuxer)
+double demuxer_get_current_time(demuxer_t *demuxer)
 {
     double get_time_ans = 0;
     sh_video_t *sh_video = demuxer->video->sh;
@@ -1488,7 +1488,7 @@ int demuxer_get_current_time(demuxer_t *demuxer)
         get_time_ans = demuxer->stream_pts;
     else if (sh_video)
         get_time_ans = sh_video->pts;
-    return (int) get_time_ans;
+    return get_time_ans;
 }
 
 int demuxer_get_percent_pos(demuxer_t *demuxer)
